weaviate/image_embedding/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
numpy>=1.26.0
weaviate-client>=4.3.0
tqdm>=4.66.0
